+++ app/Controller/RolesController.php
@@ -33,7 +33,7 @@ class RolesController extends AppController {
 	public function add() {
 		$user = $this->Auth->user();
 		$checks = array('is_add_role');
-		$result = $this->Genre->Show->User->isAuthorized($this->Genre->Show->User->Role->findById($user['role_id']), $checks);
+		$result = $this->Role->User->isAuthorized($this->Role->User->Role->findById($user['role_id']), $checks);
 		
 		if ($result['is_add_role']) {
 			if ($this->request->is('post')) {
@@ -60,81 +60,90 @@ class RolesController extends AppController {
 		if (!$role) {
 			throw new NotFoundException(__('Invalid role'));
 		}
+		
+		$user = $this->Auth->user();
+		$checks = array('is_edit_any_role');
+		$result = $this->Role->User->isAuthorized($this->Role->User->Role->findById($user['role_id']), $checks);
 	
-		if ($this->request->is('post') || $this->request->is('put')) {
-				$this->Role->id = $id;
-		  if ($this->Role->save($this->request->data)) {
-				// START TEST ACL ----------------
-				$data = $this->request->data;
-				$data = $data['Role'];
-				$this->Acl->deny($role, 'controllers');
-				$this->Acl->allow($role, 'controllers/Users/login'); // always allow any unauthenticated user to login
-				$this->Acl->allow($role, 'controllers/Users/logout'); // always allow user to logout
-				$this->Acl->allow($role, 'controllers/Users/edit'); // need to edit self
-				
-				// everything is already denied, so now we must decide what actions to allow
-				// this is a terrible way to do this, but I just want to test it right now
-				if ($data['is_add_user'])
-				{
-					$this->Acl->allow($role, 'controllers/Users/add');
+		if ($result['is_edit_any_role']) {
+			if ($this->request->is('post') || $this->request->is('put')) {
+					$this->Role->id = $id;
+			  if ($this->Role->save($this->request->data)) {
+					// START TEST ACL ----------------
+					$data = $this->request->data;
+					$data = $data['Role'];
+					$this->Acl->deny($role, 'controllers');
+					$this->Acl->allow($role, 'controllers/Users/login'); // always allow any unauthenticated user to login
+					$this->Acl->allow($role, 'controllers/Users/logout'); // always allow user to logout
+					$this->Acl->allow($role, 'controllers/Users/edit'); // need to edit self
+					
+					// everything is already denied, so now we must decide what actions to allow
+					// this is a terrible way to do this, but I just want to test it right now
+					if ($data['is_add_user'])
+					{
+						$this->Acl->allow($role, 'controllers/Users/add');
+					}
+					if ($data['is_edit_any_user'] | $data['is_edit_any_user_role'] | $data['is_edit_any_user_shows'] | $data['is_edit_any_user_status'])
+					{
+						$this->Acl->allow($role, 'controllers/Users/edit');
+					}
+					if ($data['is_add_role'])
+					{
+						$this->Acl->allow($role, 'controllers/Roles/add');
+					}
+					if ($data['is_edit_any_role'])
+					{
+						$this->Acl->allow($role, 'controllers/Roles/edit');
+					}
+					if ($data['is_add_show'])
+					{
+						$this->Acl->allow($role, 'controllers/Shows/add');
+					}
+					if ($data['is_edit_any_show'] | $data['is_edit_any_show_status'])
+					{
+						$this->Acl->allow($role, 'controllers/Shows/edit');
+					}
+					if ($data['is_add_any_episode'] | $data['is_add_authorized_episode'])
+					{
+						$this->Acl->allow($role, 'controllers/Episodes/add');
+					}
+					if ($data['is_edit_any_episode'] | $data['is_edit_authorized_episode'] | $data['is_edit_authored_episode'] |
+						$data['is_edit_any_episode_status'] | $data['is_edit_authorized_episode_status'])
+					{
+						$this->Acl->allow($role, 'controllers/Episodes/edit');
+					}
+					if ($data['is_add_genre'])
+					{
+						$this->Acl->allow($role, 'controllers/Genres/add');
+					}
+					if ($data['is_edit_any_genre'])
+					{
+						$this->Acl->allow($role, 'controllers/Genres/edit');
+					}
+					if ($data['is_add_season'])
+					{
+						$this->Acl->allow($role, 'controllers/Seasons/add');
+					}
+					if ($data['is_edit_any_season'])
+					{
+						$this->Acl->allow($role, 'controllers/Seasons/edit');
+					}
+					// this will be relevant when/if the settings are added
+				//	if ($data['is_edit_settings'])
+			//		{
+			//			$this->Acl->allow($role, 'controllers/Settings/edit');
+			//		}
+	
+					// END TEST ACL ------------------
+					$this->Session->setFlash('Your role has been updated.');
+					$this->redirect(array('action' => 'index'));
+				} else {
+					$this->Session->setFlash('Unable to update your role.');
 				}
-				if ($data['is_edit_any_user'] | $data['is_edit_any_user_role'] | $data['is_edit_any_user_shows'] | $data['is_edit_any_user_status'])
-				{
-					$this->Acl->allow($role, 'controllers/Users/edit');
-				}
-				if ($data['is_add_role'])
-				{
-					$this->Acl->allow($role, 'controllers/Roles/add');
-				}
-				if ($data['is_edit_any_role'])
-				{
-					$this->Acl->allow($role, 'controllers/Roles/edit');
-				}
-				if ($data['is_add_show'])
-				{
-					$this->Acl->allow($role, 'controllers/Shows/add');
-				}
-				if ($data['is_edit_any_show'] | $data['is_edit_any_show_status'])
-				{
-					$this->Acl->allow($role, 'controllers/Shows/edit');
-				}
-				if ($data['is_add_any_episode'] | $data['is_add_authorized_episode'])
-				{
-					$this->Acl->allow($role, 'controllers/Episodes/add');
-				}
-				if ($data['is_edit_any_episode'] | $data['is_edit_authorized_episode'] | $data['is_edit_authored_episode'] |
-					$data['is_edit_any_episode_status'] | $data['is_edit_authorized_episode_status'])
-				{
-					$this->Acl->allow($role, 'controllers/Episodes/edit');
-				}
-				if ($data['is_add_genre'])
-				{
-					$this->Acl->allow($role, 'controllers/Genres/add');
-				}
-				if ($data['is_edit_any_genre'])
-				{
-					$this->Acl->allow($role, 'controllers/Genres/edit');
-				}
-				if ($data['is_add_season'])
-				{
-					$this->Acl->allow($role, 'controllers/Seasons/add');
-				}
-				if ($data['is_edit_any_season'])
-				{
-					$this->Acl->allow($role, 'controllers/Seasons/edit');
-				}
-				// this will be relevant when/if the settings are added
-			//	if ($data['is_edit_settings'])
-		//		{
-		//			$this->Acl->allow($role, 'controllers/Settings/edit');
-		//		}
-
-				// END TEST ACL ------------------
-				$this->Session->setFlash('Your role has been updated.');
-				$this->redirect(array('action' => 'index'));
-			} else {
-				$this->Session->setFlash('Unable to update your role.');
 			}
+		} else {
+			$this->Session->setFlash('You do not have permission to edit a role.');
+			$this->redirect(array('action' => 'index'));
 		}
 	
 		if (!$this->request->data) {
@@ -147,9 +156,20 @@ class RolesController extends AppController {
 				throw new MethodNotAllowedException();
 		}
 	
-		if ($this->Role->delete($id)) {
-			$this->Session->setFlash('The role with id: ' . $id . ' has been deleted.');
-			$this->redirect(array('action' => 'index'));
+		$user = $this->Auth->user();
+		$checks = array('is_edit_any_role');
+		$result = $this->Role->User->isAuthorized($this->Role->User->Role->findById($user['role_id']), $checks);
+	
+		if ($result['is_edit_any_role']) {
+			if ($this->Role->delete($id)) {
+				$this->Session->setFlash('The role with id: ' . $id . ' has been deleted.');
+				$this->redirect(array('action' => 'index'));
+			} else {
+					$this->Session->setFlash('Unable to edit the role\'s status.');
+				}
+		} else {
+			$this->Session->setFlash('You do not have permission to edit a role\'s status.');
+			$this->redirect('/roles/index');
 		}
 	}	
 }
